attempt to fix device detection (port scanning)
See https://github.com/micro-manager/micro-manager/issues/625.  This is attempt to gracefully exit when port properties can't be read.  Currently failure happens when the port is a TCP/IP port because this port doesn't have such properties and the whole detection process is aborted (even if the device has been found).