Point and Shoot: change bleach spot detection a bit.
Now automatically assumes that a small particle is a bleached spot.
This may not be correct all the time.
Alos realized that the bleachspot is not always drawn correctly.
May need to adjust the mask so that there are clearly distinct
areas that are in or our of the bleach spot.