SutterLambda2 device adapter by Nick Anthony.
an update to the old SutterLambda adapter that now uses a hub device instead global variables. It is specific to working with their Lambda 10-3 Controller and has added functionality for Sutter's VF-5 spectral filter. For now it is named  but maybe it would be good to rename it to something else.

git-svn-id: https://valelab.ucsf.edu/svn/micromanager2/trunk@16893 d0ab736e-dc22-4aeb-8dc9-08def0aa14fd