Studio: Fix bug causing live mode to stop after circular buffer turn-over.
The display code was testing if the seq nr of the image was smaller
than that of the previous image, and would not display the image
when that happened.  I am not sure why (other than for efficiency),
but this caused the live modee window to stop updating the display
after the circular buffer was reset (i.e., full, and images inserted
at the beginning giving them seq nr. 1).  The code now only does not
display if the current and previous nr are the same.  This issue
may need to be revisited, but for now this fixes a pretty bad bug...