Build: Return to using Clojure 1.3.0.
WHen using the installer generated on the build machine, there was an error
while running an acquisition, caused by the classloader not finding org.json.JSONObject.
It is  a bit unclear where this error came from, but replacing the
MMAcquisitionENgine.jar with one generated on the local machine fixes
the issue.  I am still unsure whether or not the Clojure version is
really the cause for this error, but this seems to be the easiest
way to find out.  If this is indeed the case, upgrading Clojure
will need a thorough understanding of what is going on on the build machine.
It seems to happen on both the Mac and Windows build machines...