Studio: Fixes possible hang in Multi-Cam live mode.
The LiveModeTimer is waiting for the MultiCamLiveTask to finish.
However, the MultiCamLIveTask was waiting for an image to appear.
I added a check in the MultiCamLiveTask so that it will now exit
if the Core is not running a Sequence Acquisition.  I think that is 
a safe check, and testing shows it to overcome a hang that otherwise
occurs with longer exposure times using two cameras when stopping 
live mode.

git-svn-id: https://valelab.ucsf.edu/svn/micromanager2/trunk@17107 d0ab736e-dc22-4aeb-8dc9-08def0aa14fd