Spectral LMM5: fixing issue with certain FLICR capable versions.
Specifically seen with firmware 1.32 in a system with 6 lines.  The firmware 
reports that all lines are FLICR capable, but the max FLICR for the last 3 lines
is reported to be 0.  When MM asks for the FLICR intensity of one of those 3
lines, the controller reports an error.  
Now, that error is handled without giving up.  Also, the code tries to 
avoid asking FLICR intensities for lines tha report a 0 maximum.

git-svn-id: https://valelab.ucsf.edu/svn/micromanager2/trunk@17158 d0ab736e-dc22-4aeb-8dc9-08def0aa14fd