Studio-ScriptPanel: Correctly stores scripts in profile.
Previously, because the profile was used before the UserProfile was activated,
and after it was shutdown, settings were read and written to the default
profile most (but not all) of the time. This commit mainly changes the
order during startup, and ensures that settings will not be saved
twice during shutdown (which apparently often runs twice).