Studio: removes call animationAcknowledeDataPosition in DisplayController.
This call used DataProvider.getMaxIndices.  The code and documentation did
not make clear what this call was supposed to do.  I did not see any adverse
effects of removing the call (but not knowing what to look for made this
difficult).  Call is commented out for now.  Anyone who remembers what this
code is supposed to do, please chime in and document.