Remove 'buildcheck' script
This has not been used in a while, is unmaintained, and unlikely to be
useful in the future in this form.

For the record, the script checked for the presence of the (MM 1.4)
installer downloads, downloaded and installed them, and checked for
missing device adapters (with some hard-coded special casing).