Use US PST for nightly build dates
Until now, the time zone from which these date strings were derived was
the time zone of the computer running the builds. This is fragile,
especially on the cloud.

Fix the time zone to UTC-0800, or US Pacific Standard Time. Standard
time is used even when daylight saving time is active.

The OSX build script is no longer maintained and is left unchanged.

git-svn-id: https://valelab.ucsf.edu/svn/micromanager2/trunk@17514 d0ab736e-dc22-4aeb-8dc9-08def0aa14fd