Remove redundant autoconf macros
Before this, the micro-manager and mmCoreAndDevices repos each had an
identical copy of the m4 directory, which is a maintenance hazard.

For now, remove the copy in micro-manager and use the single copy in
mmCoreAndDevices.

If we ever move the confiuration and build of MMDevice, MMCore, and
MMCoreJ into mmCoreAndDevices, we might want to relocate some of the
macros (such as mm_ant) back into micro-manager.