Improve Autostretch when all pixels are equal
With this change, autostretch always resutls in a range with width at
least 3. When all pixels are zero, this range is 0-2, so that the image
is rendered black. When all pixels are saturated, the range consists of
the 3 largest possible values, so that the image is rendered white. When
all pixels have some non-zero, non-saturated value, the range brackets
that value, so that the image is rendered gray.

Manual adjustment of the intensity range (with autostretch disabled)
still allows ranges of width down to 2. This is intentional, because
such a setting can be useful for effectively rendering a thresholded
image. It is never useful to have a width-2 range result from
autostretch.