Keep Java source ASCII-only
The default encoding for `javac` is dependent on OS and locale. Until we
have a good mechanism to consistently pass `-encoding UTF-8` to `javac`
(such as Maven's `project.build.sourceEncoding` property), it is best to
avoid non-ASCII characters.

(Also, U+00b5 MICRO SIGN is visually indistinguishable from U+03bc GREEK
SMALL LETTER MU. It helps to be consistent in case a Unicode-unaware
search tool is used.)