Studio: display now ensures that last requested display event will actually take place.  Also, no longer change the channel color upon receiving a channelColorEvent.
Previously, the function displayImages would drop the request
to display images if it was received before painting the images
finished.  Now, this request is funneled to a ScheduledExecutor
that will display the requested images 1 second later.  If other
images are displayed in the mean time, the requst is cancelled.

The ChannelColorEvent previously caused the color of channels to
change, when the user changed the channel color in the MDA window.
That seems undesirable, so this code has been commented out and
can be removed.