Ensure Windows Ant build uses VS2019 only
Previously the build either failed, or incorrectly ran the VS2022
msbuild, when both VS2019 and VS2022 are installed.

(Using the VS2022 msbuild is a problem because the build currently
happens to fail in that case.)

This fix ensures that the VS2019 msbuild is always used when building
from Ant.

Also eliminate the impossible-to-maintain batch file syntax from this
mechanism and replace with a single PowerShell script.