Retire version.txt and simplify version setting
In the direction of switching to using CalVer for Micro-Manager
versioning, remove the 2.0.x version number everywhere (we may later
resurrect the 2.x version numbering for MMStudio API versions).

For builds using the "nightly" scripts but not explicitly setting the
version, the version now defaults to `adhoc` plus the local date
(appropriate for local test builds). Previously the date in US Pacific
Standard Time (appropriate for nightly builds) was used in this case; it
is now the invoker's responsibility to set such a date (with or without
other version information).

We also remove the distinction, in the scripts, between nightly and
release builds; for all official automated builds, the invoker must set
the whole version string (including the date, if applicable).

For plain builds not using the "nightly" scripts, no version is set, the
`MMVersion` class is not overwritten, and (on Windows) the installer
name does not contain a version (all as before). In this case,
`MMVersion` now defaults to `adhoc-build` rather than `2.0.x`.