Studio: adds coalescence to animationcontroller.
After this change, the display keeps up with fast acquisitions in real
time, i.e. it no longer runs many seconds behind and when stopping
an acquisition, it no longer displays many, may images to try to getup
to date.  There is still a slow dow in the display, but only after
20,000 images or so, and the display never comes to a complete
stand-still.